+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/filter/IntegerFilterRegistry.java
@@ -1,9 +1,6 @@
 package com.dwarfeng.fdr.impl.handler.filter;
 
-import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.annotation.JSONField;
 import com.dwarfeng.dcti.stack.bean.dto.DataInfo;
-import com.dwarfeng.fdr.impl.handler.FilterMaker;
 import com.dwarfeng.fdr.stack.bean.entity.FilterInfo;
 import com.dwarfeng.fdr.stack.bean.entity.FilteredValue;
 import com.dwarfeng.fdr.stack.exception.FilterException;
@@ -19,34 +16,45 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
-import java.util.Objects;
-
 /**
- * 正则表达式过滤器制造器。
+ * Integer过滤器注册。
  *
  * @author DwArFeng
- * @since 1.1.0
+ * @since 1.7.2
  */
 @Component
-public class RegexFilterMaker implements FilterMaker {
+public class IntegerFilterRegistry extends AbstractFilterRegistry {
 
-    public static final String SUPPORT_TYPE = "regex_filter";
+    public static final String FILTER_TYPE = "integer_filter";
 
     @Autowired
     private ApplicationContext ctx;
 
+    public IntegerFilterRegistry() {
+        super(FILTER_TYPE);
+    }
+
     @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
+    public String provideLabel() {
+        return "整型数过滤器";
+    }
+
+    @Override
+    public String provideDescription() {
+        return "如果数据值是整型数，则通过过滤。";
+    }
+
+    @Override
+    public String provideExampleContent() {
+        return "";
     }
 
     @Override
     public Filter makeFilter(FilterInfo filterInfo) throws FilterException {
         try {
-            RegexFilter filter = ctx.getBean(RegexFilter.class);
+            IntegerFilter filter = ctx.getBean(IntegerFilter.class);
             filter.setPointKey(filterInfo.getPointKey());
             filter.setFilterInfoKey(filterInfo.getKey());
-            filter.setConfig(JSON.parseObject(filterInfo.getContent(), Config.class));
             return filter;
         } catch (Exception e) {
             throw new FilterMakeException(e);
@@ -55,34 +63,33 @@ public class RegexFilterMaker implements FilterMaker {
 
     @Component
     @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public static class RegexFilter implements Filter, Bean {
+    public static class IntegerFilter implements Filter, Bean {
 
-        private static final long serialVersionUID = 970186550656361331L;
-        private static final Logger LOGGER = LoggerFactory.getLogger(RegexFilter.class);
+        private static final long serialVersionUID = -3232275174352383029L;
+        private static final Logger LOGGER = LoggerFactory.getLogger(IntegerFilter.class);
 
         private LongIdKey pointKey;
         private LongIdKey filterInfoKey;
-        private Config config;
 
-        public RegexFilter() {
+        public IntegerFilter() {
         }
 
         @Override
         public FilteredValue test(DataInfo dataInfo) throws FilterException {
             try {
                 String value = dataInfo.getValue();
-                if (!value.matches(config.getPattern())) {
-                    LOGGER.debug("测试数据值 " + dataInfo.getValue() + " 不能匹配正则表达式 " + config.getPattern()
-                            + ", 不能通过过滤...");
+                try {
+                    Integer.parseInt(value);
+                } catch (NumberFormatException e) {
+                    LOGGER.debug("测试数据值 " + dataInfo.getValue() + " 不是数字或超过整型数范围, 不能通过过滤...", e);
                     return new FilteredValue(
                             null,
                             pointKey,
                             filterInfoKey,
                             dataInfo.getHappenedDate(),
                             dataInfo.getValue(),
-                            "数据值不能匹配指定的正则表达式: " + config.getPattern()
+                            "数据值不是数字或超过整型数范围"
                     );
-
                 }
                 LOGGER.debug("测试数据值 " + dataInfo.getValue() + " 通过过滤器...");
                 return null;
@@ -107,50 +114,11 @@ public class RegexFilterMaker implements FilterMaker {
             this.filterInfoKey = filterInfoKey;
         }
 
-        public Config getConfig() {
-            return config;
-        }
-
-        public void setConfig(Config config) {
-            this.config = config;
-        }
-
         @Override
         public String toString() {
-            return "RegexFilter{" +
+            return "IntegerFilter{" +
                     "pointKey=" + pointKey +
                     ", filterInfoKey=" + filterInfoKey +
-                    ", config=" + config +
-                    '}';
-        }
-    }
-
-    public static class Config implements Bean {
-
-        private static final long serialVersionUID = -4947434410530771676L;
-
-        @JSONField(name = "pattern")
-        private String pattern;
-
-        public Config() {
-        }
-
-        public Config(String pattern) {
-            this.pattern = pattern;
-        }
-
-        public String getPattern() {
-            return pattern;
-        }
-
-        public void setPattern(String pattern) {
-            this.pattern = pattern;
-        }
-
-        @Override
-        public String toString() {
-            return "Config{" +
-                    "pattern='" + pattern + '\'' +
                     '}';
         }
     }