+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/mapper/CountMapperRegistry.java
@@ -1,7 +1,6 @@
 package com.dwarfeng.fdr.impl.handler.mapper;
 
 import com.dwarfeng.dcti.stack.bean.dto.TimedValue;
-import com.dwarfeng.fdr.impl.handler.MapperMaker;
 import com.dwarfeng.fdr.stack.exception.MapperException;
 import com.dwarfeng.fdr.stack.exception.MapperMakeException;
 import com.dwarfeng.fdr.stack.handler.Mapper;
@@ -14,27 +13,39 @@ import org.springframework.stereotype.Component;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
-import java.util.Objects;
 
 /**
- * 计数映射器制造器。
+ * 计数映射器注册。
  *
  * @author DwArFeng
- * @since 1.4.1
+ * @since 1.7.2
  */
 @Component
-public class CountMapperMaker implements MapperMaker {
+public class CountMapperRegistry extends AbstractMapperRegistry {
 
-    public static final String SUPPORT_TYPE = "count_mapper";
+    public static final String MAPPER_TYPE = "count_mapper";
 
     @Autowired
     private ApplicationContext ctx;
 
-    @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
+    public CountMapperRegistry() {
+        super(MAPPER_TYPE);
     }
 
+    @Override
+    public String provideLabel() {
+        return "计数映射器";
+    }
+
+    @Override
+    public String provideDescription() {
+        return "将输入的数据统计个数并输出，时间为统计行为发生的时间。";
+    }
+
+    @Override
+    public String provideArgsIllustrate() {
+        return "不需要任何参数";
+    }
 
     @Override
     public Mapper makeMapper(Object[] args) throws MapperException {