+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/mapper/MaxMapperRegistry.java
@@ -1,7 +1,6 @@
 package com.dwarfeng.fdr.impl.handler.mapper;
 
 import com.dwarfeng.dcti.stack.bean.dto.TimedValue;
-import com.dwarfeng.fdr.impl.handler.MapperMaker;
 import com.dwarfeng.fdr.stack.exception.MapperException;
 import com.dwarfeng.fdr.stack.exception.MapperMakeException;
 import com.dwarfeng.fdr.stack.handler.Mapper;
@@ -13,31 +12,44 @@ import org.springframework.stereotype.Component;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 
 /**
- * 最小值映射器制造器。
+ * 最大值映射器注册。
  *
  * @author DwArFeng
- * @since 1.5.3
+ * @since 1.7.2
  */
 @Component
-public class MinMapperMaker implements MapperMaker {
+public class MaxMapperRegistry extends AbstractMapperRegistry {
 
-    public static final String SUPPORT_TYPE = "min_mapper";
+    public static final String MAPPER_TYPE = "max_mapper";
 
     @Autowired
     private ApplicationContext ctx;
 
+    public MaxMapperRegistry() {
+        super(MAPPER_TYPE);
+    }
+
     @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
+    public String provideLabel() {
+        return "最大值映射器";
+    }
+
+    @Override
+    public String provideDescription() {
+        return "统计输入数据中的最大数据并输出，要求数据的全部数据均必须能被解析为double。";
+    }
+
+    @Override
+    public String provideArgsIllustrate() {
+        return "不需要任何参数";
     }
 
     @Override
     public Mapper makeMapper(Object[] args) throws MapperException {
         try {
-            return ctx.getBean(MinMapper.class);
+            return ctx.getBean(MaxMapper.class);
         } catch (Exception e) {
             throw new MapperMakeException(e);
         }
@@ -45,7 +57,7 @@ public class MinMapperMaker implements MapperMaker {
 
     @Component
     @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
-    public static class MinMapper implements Mapper {
+    public static class MaxMapper implements Mapper {
 
         @Override
         public List<TimedValue> map(List<TimedValue> timedValues) throws MapperException {
@@ -53,17 +65,17 @@ public class MinMapperMaker implements MapperMaker {
                 if (timedValues.isEmpty()) {
                     return Collections.emptyList();
                 }
-                int minIndex = 0;
-                double minValue = Double.MAX_VALUE;
+                int maxIndex = 0;
+                double maxValue = Double.MIN_VALUE;
                 for (int i = 0; i < timedValues.size(); i++) {
                     TimedValue timedValue = timedValues.get(i);
                     double v = Double.parseDouble(timedValue.getValue());
-                    if (v < minValue) {
-                        minValue = v;
-                        minIndex = i;
+                    if (v > maxValue) {
+                        maxValue = v;
+                        maxIndex = i;
                     }
                 }
-                return Collections.singletonList(timedValues.get(minIndex));
+                return Collections.singletonList(timedValues.get(maxIndex));
             } catch (Exception e) {
                 throw new MapperException(e);
             }