+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/service/TriggerInfoMaintainServiceImpl.java
@@ -96,7 +96,6 @@ public class TriggerInfoMaintainServiceImpl implements TriggerInfoMaintainServic
         return crudService.insertOrUpdate(element);
     }
 
-
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)